ratings retrieval
Continued the process of figuring out how to get the required ratings by using user_movies. Determined that the compiler errror was caused by not using set iterator *it. The expected behavior of *it is that it will be able to get the value stored at the set. I had previously only considered the map iterator.